third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-multi-line-block-end-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 2
  assert_equals: \n<item data-offset-y="68" data-expected-margin-bottom="0"></item>\noffsetTop expected 68 but got 8
[FAIL] flexbox > item 4
  assert_equals: \n<item data-offset-y="68" data-expected-margin-top="0"></item>\noffsetTop expected 68 but got 8
Harness: the test ran to completion.
